Clarify snippet placeholder structure
Now grammar implies that placeholder value should be a single `any`
node, while in most applications it can be any number of consecutive
`any` nodes. For example, as described in earlier "Placeholders"
section: `${1:another ${2:placeholder}}`.

Adding `+` quantifier to `any` seems to be more precise.

This seems to also fix the grammar for initial snippet input (which is
implied to be `any`).